% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dbx_create_file_req}
\alias{dbx_create_file_req}
\title{Create File Request via Dropbox API}
\usage{
dbx_create_file_req(folder_title, folder_path, token)
}
\arguments{
\item{folder_path}{}
}
\value{
URL string
}
\description{
Create File Request via Dropbox API
}
