% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{furthest3dates}
\alias{furthest3dates}
\title{Find furthest 3 dates from 3 lists}
\usage{
furthest3dates(dates1, dates2, dates3)
}
\arguments{
\item{dates1}{vector of dates}

\item{dates2}{vector of dates}

\item{dates3}{vector of dates}
}
\value{
list: 3 dates and the range (days)
}
\description{
Find furthest 3 dates from 3 lists
}
