% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRSS.R
\name{parseRSS}
\alias{parseRSS}
\title{Parse RSS feed, write titles & links to file}
\usage{
parseRSS(url, out_path)
}
\arguments{
\item{url}{RSS feed url}

\item{out_path}{file path to write}
}
\value{

}
\description{
Parse RSS feed, write titles & links to file
}
