% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{year_interval}
\alias{year_interval}
\title{Find number of years between two dates}
\usage{
year_interval(date1, date2)
}
\arguments{
\item{date2}{}
}
\value{
integer years (rounded down)
}
\description{
Find number of years between two dates
}
